refactor imports for file moves #531
